# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API Sistema de Préstamo de Herramientas
  version: 1.0.0
  description: API RESTful para gestionar tipos de empleados, empleados, bodegas, herramientas, préstamos y devoluciones.
servers:
  - url: http://localhost:8000/api
paths:
...
  /empleado:
    get:
      summary: Obtener todos los empleados
      responses:
        '200':
          description: Lista de empleados
    post:
      summary: Crear un nuevo empleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                apellido:
                  type: string
                cedula:
                  type: string
                id_tipo_empleado:
                  type: integer
      responses:
        '201':
          description: Empleado creado
  /empleado/{id}:
    put:
      summary: Actualizar un empleado por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                apellido:
                  type: string
                cedula:
                  type: string
                id_tipo_empleado:
                  type: integer
      responses:
        '200':
          description: Empleado actualizado
    delete:
      summary: Eliminar un empleado por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Empleado eliminado
  /bodega:
    get:
      summary: Obtener todas las bodegas
      responses:
        '200':
          description: Lista de bodegas
    post:
      summary: Crear una nueva bodega
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_bodega:
                  type: integer
                ubicacion:
                  type: string
      responses:
        '201':
          description: Bodega creada
  /bodega/{id}:
    put:
      summary: Actualizar una bodega por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ubicacion:
                  type: string
      responses:
        '200':
          description: Bodega actualizada
    delete:
      summary: Eliminar una bodega por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Bodega eliminada
...
  /devolucion:
    get:
      summary: Obtener todas las devoluciones
      responses:
        '200':
          description: Lista de devoluciones
    post:
      summary: Crear una nueva devolución
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_prestamo:
                  type: integer
                fecha_devolucion:
                  type: string
                  format: date
                hora_devolucion:
                  type: string
                  format: time
                descripcion:
                  type: string
      responses:
        '201':
          description: Devolución creada
  /devolucion/{id}:
    put:
      summary: Actualizar una devolución por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id_prestamo:
                  type: integer
                fecha_devolucion:
                  type: string
                  format: date
                hora_devolucion:
                  type: string
                  format: time
                descripcion:
                  type: string
      responses:
        '200':
          description: Devolución actualizada
    delete:
      summary: Eliminar una devolución por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Devolución eliminada